+++ src/backend/middleware/delete-jira-configuration/index.ts
@@ -1,7 +1,7 @@
-import { ActionFromSubscription, ActionSource, ActionType } from "../proto/v0/action";
-import { submitProto } from "../tracking";
-import { Subscription } from "../models";
-import getJiraClient from "../jira/client";
+import { ActionFromSubscription, ActionSource, ActionType } from "../../proto/v0/action";
+import { submitProto } from "../../tracking";
+import { Subscription } from "../../models";
+import getJiraClient from "../../jira/client";
 import { Request, Response } from "express";
 
 /**
